IS-2231: Delete old endpoints
