+++ src/main/kotlin/no/nav/syfo/application/api/ApiModule.kt
@@ -12,8 +12,6 @@ import no.nav.syfo.client.azuread.AzureAdClient
 import no.nav.syfo.client.graphapi.GraphApiClient
 import no.nav.syfo.client.wellknown.WellKnown
 import no.nav.syfo.veileder.api.registrerVeiledereApi
-import no.nav.syfo.veileder.api.registrerVeiledereApiV3
-import no.nav.syfo.veileder.api.registrerVeilederinfoApi
 import no.nav.syfo.veiledernavn.VeilederService
 
 fun Application.apiModule(
@@ -66,13 +64,7 @@ fun Application.apiModule(
         )
         registerPrometheusApi()
         authenticate(JwtIssuerType.INTERNAL_AZUREAD.name) {
-            registrerVeilederinfoApi(
-                veilederService = veilederService,
-            )
-            registrerVeiledereApi(
-                veilederService = veilederService,
-            )
-            registrerVeiledereApiV3(veilederService = veilederService)
+            registrerVeiledereApi(veilederService = veilederService)
         }
     }
 }
