+++ src/main/kotlin/no/nav/syfo/veileder/VeilederInfo.kt
@@ -1,7 +1,5 @@
 package no.nav.syfo.veileder
 
-import no.nav.syfo.veileder.api.VeilederInfoDTO
-
 data class VeilederInfo(
     val ident: String,
     val fornavn: String,
@@ -9,13 +7,3 @@ data class VeilederInfo(
     val epost: String,
     val telefonnummer: String? = null,
 )
-
-fun VeilederInfo.toVeilederinfoDTO() =
-    VeilederInfoDTO(
-        ident = this.ident,
-        navn = "${this.fornavn} ${this.etternavn}",
-        fornavn = this.fornavn,
-        etternavn = this.etternavn,
-        epost = this.epost,
-        telefonnummer = this.telefonnummer,
-    )
