--- src/main/kotlin/no/nav/syfo/veileder/api/VeilederDTO.kt
@@ -1,7 +1,0 @@
-package no.nav.syfo.veileder.api
-
-data class VeilederDTO(
-    val ident: String,
-    val fornavn: String?,
-    val etternavn: String?,
-)
